Cleanup of Engine core

+++ Engine Source/Systems/Source/EntitySystem.cpp
@@ -45,8 +45,8 @@ void EntitySystem::UpdateSystemSynchronous() CATALYST_NOEXCEPT
 void EntitySystem::ReleaseSystem() CATALYST_NOEXCEPT
 {
 	//Destroy all remaining entities.
-	for (size_t i = Entity::entities.Size() - 1; i >= 0; --i)
+	for (int64 i = Entity::entities.Size() - 1; i >= 0; --i)
 	{
-		delete Entity::entities[i];
+		delete Entity::entities[static_cast<size_t>(i)];
 	}
 }
